+++ c-implementation/tab.c
@@ -11,12 +11,15 @@
 #include <stdint.h> // uint8_t, uint32_t
 
 // TAB
-#include <tab.h>    // Header file
+#include <tab.h>    // TAB header
 
 // External handler functions
 
 extern int handle_common_data(common_data_t common_data_buff_i);
 extern int handle_bootloader_erase(void);
+extern int handle_bootloader_write_page(rx_cmd_buff_t* rx_cmd_buff);
+extern int handle_bootloader_write_page_addr32(rx_cmd_buff_t* rx_cmd_buff);
+extern int handle_bootloader_jump(void);
 extern int bootloader_active(void);
 
 // Helper functions
@@ -195,6 +198,62 @@ void write_reply(rx_cmd_buff_t* rx_cmd_buff_o, tx_cmd_buff_t* tx_cmd_buff_o) {
             tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x07);
             tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_ACK_OPCODE;
             tx_cmd_buff_o->data[PLD_START_INDEX] = BOOTLOADER_ACK_REASON_ERASED;
+          } else {
+            tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+            tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_NACK_OPCODE;
+          }
+        } else {
+          tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+          tx_cmd_buff_o->data[OPCODE_INDEX] = COMMON_NACK_OPCODE;
+        }
+        break;
+      case BOOTLOADER_WRITE_PAGE_OPCODE:
+        if(bootloader_active()) {
+          success = handle_bootloader_write_page(rx_cmd_buff_o);
+          if(success) {
+            tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x07);
+            tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_ACK_OPCODE;
+            tx_cmd_buff_o->data[PLD_START_INDEX] =
+             rx_cmd_buff_o->data[PLD_START_INDEX];
+          } else {
+            tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+            tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_NACK_OPCODE;
+          }
+        } else{
+          tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+          tx_cmd_buff_o->data[OPCODE_INDEX] = COMMON_NACK_OPCODE;
+        }
+        break;
+      case BOOTLOADER_WRITE_PAGE_ADDR32_OPCODE:
+        if(bootloader_active()) {
+          success = handle_bootloader_write_page_addr32(rx_cmd_buff_o);
+          if(success) {
+            tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x0a);
+            tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_ACK_OPCODE;
+            tx_cmd_buff_o->data[PLD_START_INDEX] =
+             rx_cmd_buff_o->data[PLD_START_INDEX];
+            tx_cmd_buff_o->data[PLD_START_INDEX+1] =
+             rx_cmd_buff_o->data[PLD_START_INDEX+1];
+            tx_cmd_buff_o->data[PLD_START_INDEX+2] =
+             rx_cmd_buff_o->data[PLD_START_INDEX+2];
+            tx_cmd_buff_o->data[PLD_START_INDEX+3] =
+             rx_cmd_buff_o->data[PLD_START_INDEX+3];
+          } else {
+            tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+            tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_NACK_OPCODE;
+          }
+        } else{
+          tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+          tx_cmd_buff_o->data[OPCODE_INDEX] = COMMON_NACK_OPCODE;
+        }
+        break;
+      case BOOTLOADER_JUMP_OPCODE:
+        if(bootloader_active()) {
+          success = handle_bootloader_jump();
+          if(success) {
+            tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x07);
+            tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_ACK_OPCODE;
+            tx_cmd_buff_o->data[PLD_START_INDEX] = BOOTLOADER_ACK_REASON_JUMPED;
           } else {
             tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
             tx_cmd_buff_o->data[OPCODE_INDEX] = BOOTLOADER_NACK_OPCODE;