+++ c-implementation/tab.h
@@ -22,19 +22,22 @@
 #define START_BYTE_1 ((uint8_t)0x69)
 
 //// Opcodes
-#define COMMON_ACK_OPCODE       ((uint8_t)0x10)
-#define COMMON_NACK_OPCODE      ((uint8_t)0xff)
-#define COMMON_DEBUG_OPCODE     ((uint8_t)0x11)
-#define COMMON_DATA_OPCODE      ((uint8_t)0x16)
-#define BOOTLOADER_ACK_OPCODE   ((uint8_t)0x01)
-#define BOOTLOADER_NACK_OPCODE  ((uint8_t)0x0f)
-#define BOOTLOADER_PING_OPCODE  ((uint8_t)0x00)
-#define BOOTLOADER_ERASE_OPCODE ((uint8_t)0x0c)
+#define COMMON_ACK_OPCODE                   ((uint8_t)0x10)
+#define COMMON_NACK_OPCODE                  ((uint8_t)0xff)
+#define COMMON_DEBUG_OPCODE                 ((uint8_t)0x11)
+#define COMMON_DATA_OPCODE                  ((uint8_t)0x16)
+#define BOOTLOADER_ACK_OPCODE               ((uint8_t)0x01)
+#define BOOTLOADER_NACK_OPCODE              ((uint8_t)0x0f)
+#define BOOTLOADER_PING_OPCODE              ((uint8_t)0x00)
+#define BOOTLOADER_ERASE_OPCODE             ((uint8_t)0x0c)
+#define BOOTLOADER_WRITE_PAGE_OPCODE        ((uint8_t)0x02)
+#define BOOTLOADER_WRITE_PAGE_ADDR32_OPCODE ((uint8_t)0x20)
+#define BOOTLOADER_JUMP_OPCODE              ((uint8_t)0x0b)
 
 //// BOOTLOADER_ACK reasons
 #define BOOTLOADER_ACK_REASON_PONG   ((uint8_t)0x00)
 #define BOOTLOADER_ACK_REASON_ERASED ((uint8_t)0x01)
-#define BOOTLOADER_ACK_REASON_JUMP   ((uint8_t)0xff)
+#define BOOTLOADER_ACK_REASON_JUMPED   ((uint8_t)0xff)
 
 //// Route Nibble IDs
 #define GND ((uint8_t)0x00)